+++ R/AFM.histogram.R
@@ -10,7 +10,7 @@
 #' @author Thomas Gredig
 #'
 #' @examples
-#' d = AFM.import(AFM.getSampleImages(type='ibw')[1])
+#' d = AFM.import(AFM.getSampleImages(type='ibw'))
 #' head(AFM.histogram(d, dataOnly=TRUE),n=20)
 #' AFM.histogram(d)
 #'